+++ modules/stocks/finnhub/settings.go
